WIP moving controlled spring mass plant
