+++ drake/systems/analysis/test/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_executable(simulator_test simulator_test.cc)
 add_executable(explicit_euler_integrator_test explicit_euler_integrator_test.cc)
 add_executable(runge_kutta2_integrator_test runge_kutta2_integrator_test.cc)
-target_link_libraries(simulator_test 
+target_link_libraries(simulator_test
   drakeSystemAnalysis drakeSystemFramework drakeCommon
         drakeSpringMassSystemPlant drakeControlledSpringMassSystemPlant
         GTest::GTest GTest::Main)
