+++ drake/systems/analysis/test/simulator_test.cc
@@ -6,7 +6,7 @@
 
 #include "drake/systems/analysis/explicit_euler_integrator.h"
 #include "drake/systems/analysis/test/my_spring_mass_system.h"
-#include "drake/systems/plants/controlled_spring_mass_system/controlled_spring_mass_system.h"
+#include "drake/systems/analysis/test/controlled_spring_mass_system/controlled_spring_mass_system.h"
 #include "drake/systems/plants/spring_mass_system/spring_mass_system.h"
 #include "gtest/gtest.h"
